fixed compiler warning
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@7526 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/pre_exodus/pre_exodus_reader.cpp
@@ -1184,7 +1184,7 @@ EXODUS::ElementBlock::ElementBlock(ElementBlock::Shape Distype, RCP<map<int,vect
   // do a sanity check
   for (map<int,vector<int> >::const_iterator elem=eleconn->begin(); elem!=eleconn->end(); ++elem)
   {
-    if (DRT::UTILS::getNumberOfElementNodes(PreShapeToDrt(Distype)) != elem->second.size())
+    if (DRT::UTILS::getNumberOfElementNodes(PreShapeToDrt(Distype)) != (int) elem->second.size())
     {
       dserror("number of read nodes does not fit the distype");
     }
